pyproject.toml
[tool.poetry]
name = "yml2dfd"
version = "0.1.0"
description = "Convert YAML files to data-flow diagrams"
authors = ["Matthew Karas <1146938+lucidguppy@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
graphviz = "^0.20"
PyYAML = "^6.0"
click = "^8.1.3"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
ipdb = "^0.13.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
